* commit '82f055e12c23475cad98f07d0fede7d20a59ba87':
  bump minor version
  Use password auth on empty clientIdName.
  Ensure that summaries have no duplicates.
  Add types for source ans sink.
  Add properties to use full qualified names.
  Change int to long

+++ api/src/main/java/com/ripstech/apiconnector2/service/application/scan/ProcessService.java
@@ -39,14 +39,14 @@ public class ProcessService extends PostGetService<Process, ProcessPost> {
 		return String.format("applications/%d/scans/%d/processes", applicationId, scanId);
 	}
 
-	public ApiResponse<Process> patch(int processId, ProcessPatch finish) {
+	public ApiResponse<Process> patch(long processId, ProcessPatch finish) {
 		return new ApiResponse<>(getTarget(PATCH)
 				                         .appendPath(processId)
 				                         .setJsonBody(finish),
 		                         getGenericType());
 	}
 
-	public ApiResponse<Process> patch(int processId, OffsetDateTime finish) {
+	public ApiResponse<Process> patch(long processId, OffsetDateTime finish) {
 		return patch(processId, ProcessSend.createPatch().setFinish(finish));
 	}
